dbapi_service/requirements.txt
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
distlib==0.3.6
filelock==3.8.0
Flask==2.2.2
# Flask-MySQLdb==1.0.1
# Flask-SQLAlchemy==2.5.1
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.30
# mariadb-connector-python==1.0.11
# mysqlclient==2.1.1
# numpy==1.23.3
# pandas==1.4.4
platformdirs==2.5.2
protobuf==3.20.1
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
SQLAlchemy==1.4.41
tqdm==4.64.1
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
mariadb==1.0.11
